feat: Add chat instance initialization and message listener
The Chat model has been updated to include the initialization of chat instances. A listener for new chat messages has also been added. Once a new message is pushed to the chat, the listener will trigger and the view will automatically scroll to the new message. In addition, the Http.get method and new endpoint have been added to the Http model to handle chat instance creation.